legal-ai-assistant/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-ratelimit==4.1.0
celery==5.3.4
redis==5.0.1

# LLM inference
llama-cpp-python==0.2.20

# Fine-tuning and ML
torch==2.1.0
transformers==4.35.2
peft==0.7.0
bitsandbytes==0.41.3
accelerate==0.25.0
datasets==2.15.0
sentencepiece==0.1.99
protobuf==4.25.1

# Utilities
python-dotenv==1.0.0
PyPDF2==3.0.1
python-magic==0.4.27
tqdm==4.66.1
numpy==1.26.2
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
